test-with-pom - move sequences into _lib/sequences
