+++ e2e/user/user-can-logout.spec.ts
@@ -1,6 +1,6 @@
 import { test, expect } from "@playwright/test";
 import { loginKnownUser, navbarOf } from "../_lib/pom/init";
-import { doLogoutSequence } from "../_lib/pom/sequences/logout-sequence";
+import { doLogoutSequence } from "../_lib/sequences/logout-sequence";
 
 test("user can logout", async ({ page }) => {
   const { knownUser, pomPage } = await loginKnownUser(page);
